# Optimize Response#parse_headers