@toland 2015-08-15
Thanks for the changes and a very thorough pull request.